fix: Bug fixes to CLI and RayChannel (#67)
# Summary
Various bug fixes.

# Changes
* Resolves an issue where occasionally testing the `RayChannel` fails.
This is because the test was using `asyncio.gather` to receive
consecutive values from the channel, which preserves order in the local
case but is not guaranteed when using Ray async actors.
* Also improves the CLI by adding the location of the model YAML file to
`sys.path`. This is necessary because `sys.path` does not contain the
current working directory by default (see
[here](https://docs.python.org/3/library/sys_path_init.html#the-initialization-of-the-sys-path-module-search-path)).
Making this change ensures that any Python files defining components
alongside the model YAML will be importable.
* Fixes channel close in `DataReader` component that was causing models
to hang.