feat: ZMQ pubsub connector (#71)
# Summary

This PR supports #55 by adding in new classes which enable use of ZMQ
sockets for communication in a pubsub pattern. With regards to the
specific requirements of gracefully stopping event based models, the
pubsub comminucation pattern will allow all components in a process to
publish and subscribe to a planned builtin "StopEvent" event. When any
component emits this event all others will receive it and subsequently
exit gracefully.

To support a wider range of communication patterns and channel types,
some significant changes have been made to the lower level classes in
the code which handle inter-component communication. This is in
recognition of the need, in the general case, to create `Channel`s on
the os process which consumes them, avoiding the need to serialise and
deserialise objects - such as ZMQ sockets - for which it is either
unsafe or infeasible to share between os processes (which may or may not
be on the same machine).

# Changes
- Changes the high level API for creating channels by modifying the
roles and relationships between some core classes:
- `ChannelBuilder` is now `ConnectorBuilder` to move responsibility of
`Channel` creation down to `Connector` to ensure that `Channel`s are
created in the os process which consumes them.
- `Connector` is now an abstract base class with concrete
implementations (`AsyncioConnector`, `RayConnector`, `ZMQConnector` etc)
which build and connect specific `Channel` implementations.
- Changes the values of the `ConnectorMode` Enum to `pipeline`
(one-in-one-out work queue type message distribution) and `pubsub`
(one-to-many broadcast type communication).
- Introduces private `_ZMQConnector`, `_ZMQPipelineConnector`,
`_ZMQPubsubConnector`, and `_ZMQPubsubConnectorProxy` which implement
different ZMQ messaging patterns
using appropriate ZMQ sockets and configurations.
- Introduces a `ZMQProxy` class inheriting from
`multiprocessing.Process` which runs a proxy for ZMQ socket connections
with libzmq in a separate OS process. This is gated behind a feature
flag `PLUGBOARD_FLAGS_ZMQ_PUBSUB_PROXY`
- Introduces a public `ZMQConnector` which owns an internal private
`_ZMQConnector` implementation which is selected based on the specified
connector mode.
- Introduces a dependency injection container using
[`that-depends`](https://github.com/modern-python/that-depends) DI
framework. In particular this manages the `ZMQProxy` singleton
lifecycle.