fix: Ray serialisation issues (#92)
# Summary
Closes #84 and #85 with bug fixes for the Ray integration. 

# Changes
* Adds `exports` attribute to `Component`, allowing users to specify
which attributes they want returned to the Ray driver at the end of a
run. This prevents serialisation issues with the existing code.
* Adds `is_on_ray_worker` utility to detect whether running on a Ray
worker.
* Skip the `__init_subclass__` method on `Component` when inside a Ray
worker. This means that `ComponentRegistry` won't be updated, but that
should not be an issue on the Ray worker. Required for components
defined in the `__main__` module owing to
https://github.com/ray-project/ray/issues/42823.