feat: Component decorator (#169)
# Summary
This PR adds a `plugboard.component.utils` module which includes a
`component` decorator. When applied to a sync or async function a
`Component` class is created dynamically which calls the decorated
function in the `Component.step` method. The function arguments should
match those specified as the `input` argument to the `component`
decorator, and the return value must be a dictionary with keys matching
the `output` argument to the `component` decorator. Calling the
decorated function results in a new `ComponentDecoratorHelper` instance
which preserves the original behaviour of the wrapped function on call,
and provides a `.component` method to instantiate an instance of the
dynamically created `FuncComponent` class.

# Changes
- Adds a `plugboard.component.utils` module which includes a `component`
decorator
- Adds tests for decorated function components.
- Fixes a bug in the RabbitMQ connector in pipeline mode which was
distributing messages for a given output field amongst all receivers.